对And64InlineHook.cpp的代码做了一丁点修改，并给出了一个inline-hook用例
