+++ test_example/Android.mk
